requirements-dev.txt
# Testing libraries
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Code quality tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Coverage and reporting
coverage>=7.0.0

# Mock and testing utilities
responses>=0.23.0
fakeredis>=2.10.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development utilities
pre-commit>=3.0.0
ipython>=8.0.0
jupyter>=1.0.0

# Base requirements (should match requirements.txt)
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0